refactor(#3255): merge api pre and post
